issue#8 move KitCard to kits/components/KitCard
